[v0.90][WP-14] Split cognitive state artifact responsibilities (Closes #2052)
